pyproject.toml
[build-system]
requires = [ "poetry-core>=1.9.1",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-auth-logger"
version = "2.2.1"
description = "A tiny project to log login attempts. Log them only to standard logging - not the database."
authors = ["thelab <thelabdev@thelab.co>"]
readme = "README.rst"
homepage = "https://gitlab.com/thelabnyc/django-auth-logger"
repository = "https://gitlab.com/thelabnyc/django-auth-logger"
license = "ISC"

[[tool.poetry.packages]]
include = "auth_logger"
from = "src"

[tool.poetry.dependencies]
python = "^3.10"
Django = ">=4.2"
django-ipware = ">=7.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = ">=7.1.1"
freezegun = ">=1.5.1"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.1"}

[tool.commitizen]
name = "cz_conventional_commits"
annotated_tag = true
gpg_sign = true
tag_format = "v$version"
update_changelog_on_bump = true
changelog_merge_prerelease = true
version_provider = "poetry"
version_scheme = "pep440"
version_files = [
    "pyproject.toml:version",
]
pre_bump_hooks = [
    "pre-commit run --all-files || true",
]
post_bump_hooks = [
    "git push origin master $CZ_POST_CURRENT_TAG_VERSION"
]
